moved chord events
