+++ README.md
@@ -10,3 +10,11 @@ py midi_server.py
 cd chord-game
 npm run dev
 ```
+
+```
+TODO:
+	- display broken chord sequence
+	- rewrite entire code in terms of Chord / BrokenChord Data structures
+	- create a way of making and storing motifs
+	- make the motif game
+```
